Backend para botones sociales

+++ backend/controllers/users.js
@@ -17,14 +17,18 @@ export const loginUser = async (req, res) => {
         // Verificar si los campos estan vacios
         if (!email || !password) return res.status(400).json({ message: 'Todos los campos son obligatorios' });
         // buscar el usuario por su correo electrónico
-        const user = await Users.findOne({ email });
+        const user = await Users.findOne({ email }).select('+password');
         // validamos que el usuario exista en nuestra base da datos
         if (!user) return res.status(401).json({ message: 'Usuario o contraseña incorrectos' });
-        // Si se registró con el formulario
-        if (user.provider === 'local') {
+        // Si se registró con el formulario previamente
+        if (user.password) {
             // Comparar contraseña hasheada con la contraseña proporcionada
             const passwordValid = await bcrypt.compare(password, user.password);
             if (!passwordValid) return res.status(400).json({ message: "Contraseña incorrecta" });
+        } else{
+            return res.status(400).json({ 
+                message: "Este correo está vinculado a una cuenta de Google o Facebook. Por favor, inicia sesión correctamente." 
+            });
         }
         // Validar verificación
         if (!user.isVerified) {
@@ -34,8 +38,7 @@ export const loginUser = async (req, res) => {
                 email: user.email // Útil para que el frontend sepa a dónde reenviar el código
             });
         }
-
-
+        // Creamos el token de esta sesion
         const token = jwt.sign(
             { id: user._id, role: user.role || 'user' },
             process.env.JWT_SECRET || 'clave_secreta_provisional',
@@ -64,7 +67,7 @@ export const editUser = async (req, res) => {
         if (!expressions.name.test(name)) return res.status(400).json({ message: "Escribe un nombre o un apellido válido." });
 
         // actualizamos el nombre
-        const updatedUser = await Users.findByIdAndUpdate(req.user.id, { $set: { name: name } }, { returnDocument: 'after', runValidators: true }).select("-password");
+        const updatedUser = await Users.findByIdAndUpdate(req.user.id, { $set: { name: name } }, { returnDocument: 'after', runValidators: true });
 
         // si no se encuentra el usuario, no se actualiza y se envía un mensaje
         if (!updatedUser) return res.status(401).json({ message: "No se encontró el usuario para actualizar" });
@@ -79,12 +82,20 @@ export const editUser = async (req, res) => {
 // eliminar la cuenta
 export const deleteUser = async (req, res) => {
     try {
-        // buscamos el usuario
-        const deletedUser = await Users.findByIdAndDelete(req.user.id);
-        // validamos que exista el usuario
-        if (!deletedUser) return res.status(404).json({ message: "Usuario no encontrado" })
-        // enviamos un mensajito a nuestro usuario desertado
-        res.status(200).json({ message: `Se ha eliminado la cuentica de ${deletedUser.email}` })
+        // extraemos la contraseña para validarla y eliminar la cuenta
+        const { password } = req.body;
+        // validamos si hay
+        if(!password) return res.status(400).json({ message: "La contraseña es requerida" });
+        // buscamos el usuario por su id y validamos que exista el usuario
+        const user = await Users.findById(req.user.id).select('+password');
+        if (!user) return res.status(404).json({ message: "Usuario no encontrado" });
+        // validamos la contraseña para eliminar la cuenta
+        const isMatch = await bcrypt.compare(password, user.password);
+        if (!isMatch) return res.status(401).json({ message: "Contraseña ingresada no es correcta" });
+        // si todo esta bien, eliminamos la cuenta
+        await Users.findByIdAndDelete(req.user.id);
+        // enviamos un mensajito de feedback
+        res.status(200).json({ message: `Se ha eliminado la cuentica de ${user.email}` });
     } catch (error) {
         res.status(500).json({ message: 'Error al eliminar cuenta', error: error.message });
     }
